+++ src/pages/Billboard/BillboardHeader/PopoverTitle.tsx
@@ -1,57 +1,41 @@
+import { CloseOutlined, LeftOutlined } from "@ant-design/icons";
+import { Button } from "antd";
 import { PopoverTitleProps } from "@/interfaces/boards";
 import { PopoverTitleStyle } from "./style";
-import { Button } from "antd";
-import { CloseOutlined, LeftOutlined } from "@ant-design/icons";
 
 const PopoverTitle: React.FC<PopoverTitleProps> = ({
   headerState,
   setHeaderState,
   setOpenPopover,
 }) => {
-  const handleClick = () => {
-    setOpenPopover(false);
-    setHeaderState("MENU");
-  };
-  const previousClick = () => {
-    setHeaderState("MENU");
-  };
+  const titleText = {
+    MENU: "選單",
+    USER: "查看看板管理員",
+    SETTING: "設定",
+    LABEL: "標籤",
+  }[headerState];
 
   return (
     <PopoverTitleStyle>
-      {headerState === "MENU" && (
-        <>
-          選單
-          <Button
-            size="small"
-            type="text"
-            icon={
-              <CloseOutlined
-                style={{
-                  color: "var(--gray66)",
-                }}
-              />
-            }
-            style={{ position: "absolute", right: 3 }}
-            onClick={handleClick}
-          />
-        </>
-      )}
-      {headerState === "USER" && (
-        <>
-          查看看板管理員
-          <Button
-            size="small"
-            type="text"
-            icon={
-              <CloseOutlined
-                style={{
-                  color: "var(--gray66)",
-                }}
-              />
-            }
-            style={{ position: "absolute", right: 3 }}
-            onClick={handleClick}
-          />
+      <>
+        {titleText}
+        <Button
+          size="small"
+          type="text"
+          icon={
+            <CloseOutlined
+              style={{
+                color: "var(--gray66)",
+              }}
+            />
+          }
+          style={{ position: "absolute", right: 3 }}
+          onClick={() => {
+            setOpenPopover(false);
+            setHeaderState("MENU");
+          }}
+        />
+        {headerState !== "MENU" && (
           <Button
             size="small"
             type="text"
@@ -61,68 +45,10 @@ const PopoverTitle: React.FC<PopoverTitleProps> = ({
                 style={{ color: "var(--gray66)", fontSize: "16px" }}
               />
             }
-            onClick={previousClick}
+            onClick={() => setHeaderState("MENU")}
           />
-        </>
-      )}
-      {headerState === "SETTING" && (
-        <>
-          設定
-          <Button
-            size="small"
-            type="text"
-            icon={
-              <CloseOutlined
-                style={{
-                  color: "var(--gray66)",
-                }}
-              />
-            }
-            style={{ position: "absolute", right: 3 }}
-            onClick={handleClick}
-          />
-          <Button
-            size="small"
-            type="text"
-            style={{ position: "absolute", left: -2, top: 2 }}
-            icon={
-              <LeftOutlined
-                style={{ color: "var(--gray66)", fontSize: "16px" }}
-              />
-            }
-            onClick={previousClick}
-          />
-        </>
-      )}
-      {headerState === "LABEL" && (
-        <>
-          標籤
-          <Button
-            size="small"
-            type="text"
-            icon={
-              <CloseOutlined
-                style={{
-                  color: "var(--gray66)",
-                }}
-              />
-            }
-            style={{ position: "absolute", right: 3 }}
-            onClick={handleClick}
-          />
-          <Button
-            size="small"
-            type="text"
-            style={{ position: "absolute", left: -2, top: 2 }}
-            icon={
-              <LeftOutlined
-                style={{ color: "var(--gray66)", fontSize: "16px" }}
-              />
-            }
-            onClick={previousClick}
-          />
-        </>
-      )}
+        )}
+      </>
     </PopoverTitleStyle>
   );
 };
