--- src/pages/Billboard/BillboardHeader/constant.ts
@@ -1,11 +1,0 @@
-export const colorList = [
-  { color: "#FBF8CC" },
-  { color: "#FDE4CF" },
-  { color: "#FFCFD2" },
-  { color: "#F1C0E8" },
-  { color: "#CFBAF0" },
-  { color: "#A3C4F3" },
-  { color: "#90DBF4" },
-  { color: "#8EECF5" },
-  { color: "#98F5E1" },
-];
