+++ src/pages/Billboard/BillboardHeader/index.tsx
@@ -13,15 +13,9 @@ import PopoverContent from "./PopoverContent";
 import AddMember from "@/components/Modal/AddMember";
 
 const BillboardHeader: React.FC<BillboardHeaderProps> = ({
-  name,
-  member,
-  boardInviteLink,
-  orgId,
+  board,
+
   callGetBoardApi,
-  boardId,
-  image,
-  permission,
-  closed,
 }) => {
   const [openInvite, setOpenInvite] = useState(false);
   const [openPopover, setOpenPopover] = useState(false);
@@ -40,13 +34,13 @@ const BillboardHeader: React.FC<BillboardHeaderProps> = ({
                   #0083ff 1.48%,
                   rgba(128, 0, 255, 0.86) 100%
                 )`}
-          background-image={image && `url(${image})`}
+          background-image={board?.image && `url(${board.image})`}
         />
-        <p style={{ marginLeft: "16px" }}>{name}</p>
+        <p style={{ marginLeft: "16px" }}>{board?.name}</p>
       </div>
       <div className="right-head">
         <Avatar.Group>
-          {member?.map(({ userId: { avatar, name, _id } }) => (
+          {board?.member?.map(({ userId: { avatar, name, _id } }) => (
             <Tooltip placement="top" title={name} key={_id}>
               <Avatar src={avatar} />
             </Tooltip>
@@ -77,14 +71,8 @@ const BillboardHeader: React.FC<BillboardHeaderProps> = ({
             <PopoverContent
               headerState={headerState}
               setHeaderState={setHeaderState}
-              name={name}
-              member={member}
-              orgId={orgId}
+              board={board}
               callGetBoardApi={callGetBoardApi}
-              boardId={boardId || ""}
-              permission={permission || ""}
-              closed={closed}
-              image={image || ""}
             />
           }
           trigger="click"
@@ -104,8 +92,8 @@ const BillboardHeader: React.FC<BillboardHeaderProps> = ({
       <AddMember
         open={openInvite}
         setOpen={setOpenInvite}
-        member={member}
-        boardInviteLink={boardInviteLink}
+        member={board?.member || []}
+        boardInviteLink={board?.inviteLink || ""}
       />
     </BillboardHeaderCss>
   );
