+++ src/utils/constant.ts
@@ -30,3 +30,15 @@ export const errorMessage: { [x: number]: string } = {
 };
 
 export const POSITION_GAP = 65535;
+
+export const colorList = [
+  { color: "#FBF8CC" },
+  { color: "#FDE4CF" },
+  { color: "#FFCFD2" },
+  { color: "#F1C0E8" },
+  { color: "#CFBAF0" },
+  { color: "#A3C4F3" },
+  { color: "#90DBF4" },
+  { color: "#8EECF5" },
+  { color: "#98F5E1" },
+];
